requirements.txt
fastapi
uvicorn
jinja2
gitpython
sphinx
sphinx-autodoc-typehints
markdown 